Change error handling in the model
In the ACL2 model, functions that might produce an error return an error
flag that is checked by the caller, and either passed on by auxiliary
functions or reported by instructions using clauses like `((when flg)
(!!fault-fresh ...))` in a `b*` expression (which will, if the flag is
set, report the error and stop the execution of the instruction).

We have so far translated the generation and passing around of error
flags with the `option(string)` type, and the reporting of errors via
throwing Sail exceptions.

This patch pushes the throwing of exceptions down to the points in
auxiliary functions where errors are generated, and detects and removes
(with some hackery in the translator) the passing around and checking of
error flags, which leads to much less verbose generated Sail code.