Constrain types of some variables and functions
Avoids gratuitous extension of bitvectors when we already know that we
will only be using the lower bits.

In particular, set the return type of `add_to_iptr` to bits(48), which
is safe because it internally checks that the result will fit into 48
bits.

Also fix some bugs that caused type overrides for variables to not be
picked up in some corner cases.